+++ index.js
@@ -1,6 +1,8 @@
-const app = require('./app');
+const createServer = require('./server');
 const db = require('./db/connection');
 const PORT = process.env.PORT || 3000;
+
+const app = createServer();
 
 db.on('error', console.error.bind(console, 'MongoDB connection error:'));
